fix(auth): moved error constructors in core package
