+++ packages/auth/src/core/strategies/minitailor.ts
@@ -1,5 +1,4 @@
-import { AbstractStrategy } from "./abstract";
-import { paramsError } from "@server/middleware/callback";
+import { AbstractStrategy, paramsError } from "@core/strategies/abstract";
 import { Config } from "@core/config";
 import { callbackByStrategy } from "@core/path";
