+++ packages/auth/src/server/middleware/callback.ts
@@ -2,20 +2,7 @@ import { NextResponse } from "next/server";
 import { RouteHandler } from "@server/middleware";
 import { ErrorResponse, Session } from "@core/types";
 import { Config } from "@core/config";
-
-export class CallbackError extends Error {
-  constructor(
-    readonly name: string,
-    readonly message: string,
-  ) {
-    super(message);
-  }
-}
-
-export const paramsError = () =>
-  new CallbackError("invalid-params", "code and redirectURI should be filled");
-export const exchangeError = (reason: string) =>
-  new CallbackError("failed-exchange", reason);
+import { exchangeError } from "@core/strategies/abstract";
 
 export const callbackHandler: RouteHandler = async ({
   request,
